move calendar button to proper location
