+++ app/components/booking/form.tsx
@@ -8,6 +8,12 @@ import { useAtom } from "jotai";
 import { useZorm } from "react-zorm";
 import { z } from "zod";
 import { updateDynamicTitleAtom } from "~/atoms/dynamic-title-atom";
+import {
+  Tooltip,
+  TooltipContent,
+  TooltipProvider,
+  TooltipTrigger,
+} from "~/components/shared/tooltip";
 import { useBookingStatus } from "~/hooks/use-booking-status";
 import { useUserIsSelfService } from "~/hooks/user-user-is-self-service";
 import type { BookingWithCustodians } from "~/routes/_layout+/bookings";
@@ -21,6 +27,7 @@ import { AbsolutePositionedHeaderActions } from "../layout/header/absolute-posit
 import { Button } from "../shared";
 import { Card } from "../shared/card";
 import { ControlledActionButton } from "../shared/controlled-action-button";
+
 /**
  * Important note is that the fields are only valudated when they are not disabled
  */
@@ -300,6 +307,7 @@ export function BookingForm({
                   assets during the duration of the booking period.
                 </p>
               </Card>
+              <AddToCalendar />
             </div>
           </div>
         </div>
@@ -312,3 +320,30 @@ export function BookingForm({
     </div>
   );
 }
+
+const AddToCalendar = () => {
+  const navigation = useNavigation();
+  const disabled = isFormProcessing(navigation.state);
+
+  return (
+    <TooltipProvider delayDuration={100}>
+      <Tooltip>
+        <TooltipTrigger asChild>
+          <Button
+            to={`cal.ics`}
+            download={true}
+            reloadDocument={true}
+            disabled={disabled}
+            variant="secondary"
+            icon="calendar"
+          >
+            Add to calendar
+          </Button>
+        </TooltipTrigger>
+        <TooltipContent side="bottom">
+          <p className="text-xs">Download this booking as a calendar event</p>
+        </TooltipContent>
+      </Tooltip>
+    </TooltipProvider>
+  );
+};
